Add files via upload
A static Web page is a page that is built using HTML code and features the same presentation and content, regardless of user identity or other factors. Static Web pages are easier to code and assemble than dynamic Web pages, which may feature customizable content according to a user's identity or other factors.